+++ internal/pkg/oci/defaults.go
@@ -1,8 +1,9 @@
 package oci
 
 import (
-	warewulfconf "github.com/warewulf/warewulf/internal/pkg/config"
 	"path/filepath"
+
+	warewulfconf "github.com/warewulf/warewulf/internal/pkg/config"
 )
 
 var defaultCachePath = filepath.Join(warewulfconf.Get().Warewulf.DataStore, "/container-cache/oci/")