+++ internal/pkg/warewulfd/provision.go
@@ -73,21 +73,21 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 	// TODO: when module version is upgraded to go1.18, should be 'any' type
 	var tmpl_data interface{}
 
-	node, err := GetNodeOrSetDiscoverable(rinfo.hwaddr)
-	if err != nil {
+	remoteNode, err := GetNodeOrSetDiscoverable(rinfo.hwaddr)
+	if err != nil && err != node.ErrNoUnconfigured {
 		wwlog.ErrorExc(err, "")
 		w.WriteHeader(http.StatusServiceUnavailable)
 		return
 	}
 
-	if node.AssetKey.Defined() && node.AssetKey.Get() != rinfo.assetkey {
+	if remoteNode.AssetKey != "" && remoteNode.AssetKey != rinfo.assetkey {
 		w.WriteHeader(http.StatusUnauthorized)
-		wwlog.Denied("Incorrect asset key for node: %s", node.Id.Get())
-		updateStatus(node.Id.Get(), status_stage, "BAD_ASSET", rinfo.ipaddr)
+		wwlog.Denied("Incorrect asset key for node: %s", remoteNode.Id())
+		updateStatus(remoteNode.Id(), status_stage, "BAD_ASSET", rinfo.ipaddr)
 		return
 	}
 
-	if !node.Id.Defined() {
+	if !remoteNode.Valid() {
 		wwlog.Error("%s (unknown/unconfigured node)", rinfo.hwaddr)
 		if rinfo.stage == "ipxe" {
 			stage_file = path.Join(conf.Paths.Sysconfdir, "/warewulf/ipxe/unconfigured.ipxe")
@@ -96,45 +96,44 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 		}
 
 	} else if rinfo.stage == "ipxe" {
-		stage_file = path.Join(conf.Paths.Sysconfdir, "warewulf/ipxe/"+node.Ipxe.Get()+".ipxe")
-		tstruct := overlay.InitStruct(&node)
+		stage_file = path.Join(conf.Paths.Sysconfdir, "warewulf/ipxe/"+remoteNode.Ipxe+".ipxe")
 		tmpl_data = templateVars{
-			Id:             node.Id.Get(),
-			Cluster:        node.ClusterName.Get(),
-			Fqdn:           node.Id.Get(),
+			Id:             remoteNode.Id(),
+			Cluster:        remoteNode.ClusterName,
+			Fqdn:           remoteNode.Id(),
 			Ipaddr:         conf.Ipaddr,
 			Port:           strconv.Itoa(conf.Warewulf.Port),
-			Hostname:       node.Id.Get(),
+			Hostname:       remoteNode.Id(),
 			Hwaddr:         rinfo.hwaddr,
-			ContainerName:  node.ContainerName.Get(),
-			KernelArgs:     node.Kernel.Args.Get(),
-			KernelOverride: node.Kernel.Override.Get(),
-			NetDevs:        tstruct.NetDevs,
-			Tags:           tstruct.Tags}
+			ContainerName:  remoteNode.ContainerName,
+			KernelArgs:     remoteNode.Kernel.Args,
+			KernelOverride: remoteNode.Kernel.Override,
+			NetDevs:        remoteNode.NetDevs,
+			Tags:           remoteNode.Tags}
 	} else if rinfo.stage == "kernel" {
-		if node.Kernel.Override.Defined() {
-			stage_file = kernel.KernelImage(node.Kernel.Override.Get())
-		} else if node.ContainerName.Defined() {
-			stage_file, _, err = kernel.FindKernel(container.RootFsDir(node.ContainerName.Get()))
+		if remoteNode.Kernel.Override != "" {
+			stage_file = kernel.KernelImage(remoteNode.Kernel.Override)
+		} else if remoteNode.ContainerName != "" {
+			stage_file, _, err = kernel.FindKernel(container.RootFsDir(remoteNode.ContainerName))
 			if err != nil {
-				wwlog.Error("No kernel found for container %s: %s", node.ContainerName.Get(), err)
+				wwlog.Error("No kernel found for container %s: %s", remoteNode.ContainerName, err)
 			}
 		} else {
-			wwlog.Warn("No kernel version set for node %s", node.Id.Get())
+			wwlog.Warn("No kernel version set for node %s", remoteNode.Id())
 		}
 
 	} else if rinfo.stage == "kmods" {
-		if node.Kernel.Override.Defined() {
-			stage_file = kernel.KmodsImage(node.Kernel.Override.Get())
+		if remoteNode.Kernel.Override != "" {
+			stage_file = kernel.KmodsImage(remoteNode.Kernel.Override)
 		} else {
-			wwlog.Warn("No kernel override modules set for node %s", node.Id.Get())
+			wwlog.Warn("No kernel override modules set for node %s", remoteNode.Id())
 		}
 
 	} else if rinfo.stage == "container" {
-		if node.ContainerName.Defined() {
-			stage_file = container.ImageFile(node.ContainerName.Get())
+		if remoteNode.ContainerName != "" {
+			stage_file = container.ImageFile(remoteNode.ContainerName)
 		} else {
-			wwlog.Warn("No container set for node %s", node.Id.Get())
+			wwlog.Warn("No container set for node %s", remoteNode.Id())
 		}
 
 	} else if rinfo.stage == "system" || rinfo.stage == "runtime" {
@@ -147,7 +146,7 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 			context = rinfo.stage
 		}
 		stage_file, err = getOverlayFile(
-			node,
+			remoteNode,
 			context,
 			request_overlays,
 			conf.Warewulf.AutobuildOverlays)
@@ -164,7 +163,7 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 		}
 	} else if rinfo.stage == "efiboot" {
 		wwlog.Debug("requested method: %s", req.Method)
-		containerName := node.ContainerName.Get()
+		containerName := remoteNode.ContainerName
 		switch rinfo.efifile {
 		case "shim.efi":
 			stage_file = container.ShimFind(containerName)
@@ -182,20 +181,17 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 			}
 		case "grub.cfg":
 			stage_file = path.Join(conf.Paths.Sysconfdir, "warewulf/grub/grub.cfg.ww")
-			tstruct := overlay.InitStruct(&node)
 			tmpl_data = templateVars{
-				Id:             node.Id.Get(),
-				Cluster:        node.ClusterName.Get(),
-				Fqdn:           node.Id.Get(),
+				Id:             remoteNode.Id(),
+				Cluster:        remoteNode.ClusterName,
+				Fqdn:           remoteNode.Id(),
 				Ipaddr:         conf.Ipaddr,
 				Port:           strconv.Itoa(conf.Warewulf.Port),
-				Hostname:       node.Id.Get(),
+				Hostname:       remoteNode.Id(),
 				Hwaddr:         rinfo.hwaddr,
-				ContainerName:  node.ContainerName.Get(),
-				KernelArgs:     node.Kernel.Args.Get(),
-				KernelOverride: node.Kernel.Override.Get(),
-				NetDevs:        tstruct.NetDevs,
-				Tags:           tstruct.Tags}
+				ContainerName:  remoteNode.ContainerName,
+				KernelArgs:     remoteNode.Kernel.Args,
+				KernelOverride: remoteNode.Kernel.Override}
 			if stage_file == "" {
 				wwlog.ErrorExc(fmt.Errorf("could't find grub.cfg template"), containerName)
 				w.WriteHeader(http.StatusNotFound)
@@ -205,36 +201,36 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 			wwlog.ErrorExc(fmt.Errorf("could't find efiboot file: %s", rinfo.efifile), "")
 		}
 	} else if rinfo.stage == "shim" {
-		if node.ContainerName.Defined() {
-			stage_file = container.ShimFind(node.ContainerName.Get())
+		if remoteNode.ContainerName != "" {
+			stage_file = container.ShimFind(remoteNode.ContainerName)
 
 			if stage_file == "" {
-				wwlog.Error("No kernel found for container %s", node.ContainerName.Get())
+				wwlog.Error("No kernel found for container %s", remoteNode.ContainerName)
 			}
 		} else {
-			wwlog.Warn("No container set for this %s", node.Id.Get())
+			wwlog.Warn("No container set for this %s", remoteNode.Id())
 		}
 	} else if rinfo.stage == "grub" {
-		if node.ContainerName.Defined() {
-			stage_file = container.GrubFind(node.ContainerName.Get())
+		if remoteNode.ContainerName != "" {
+			stage_file = container.GrubFind(remoteNode.ContainerName)
 			if stage_file == "" {
-				wwlog.Error("No grub found for container %s", node.ContainerName.Get())
+				wwlog.Error("No grub found for container %s", remoteNode.ContainerName)
 			}
 		} else {
-			wwlog.Warn("No conainer set for node %s", node.Id.Get())
+			wwlog.Warn("No conainer set for node %s", remoteNode.Id())
 		}
 	} else if rinfo.stage == "initramfs" {
-		if node.ContainerName.Defined() {
-			_, kver, err := kernel.FindKernel(container.RootFsDir(node.ContainerName.Get()))
+		if remoteNode.ContainerName != "" {
+			_, kver, err := kernel.FindKernel(container.RootFsDir(remoteNode.ContainerName))
 			if err != nil {
-				wwlog.Error("No kernel found for initramfs for container %s: %s", node.ContainerName.Get(), err)
+				wwlog.Error("No kernel found for initramfs for container %s: %s", remoteNode.ContainerName, err)
 			}
-			stage_file, err = container.InitramfsBootPath(node.ContainerName.Get(), kver)
+			stage_file, err = container.InitramfsBootPath(remoteNode.ContainerName, kver)
 			if err != nil {
-				wwlog.Error("No initramfs found for container %s: %s", node.ContainerName.Get(), err)
+				wwlog.Error("No initramfs found for container %s: %s", remoteNode.ContainerName, err)
 			}
 		} else {
-			wwlog.Warn("No container set for node %s", node.Id.Get())
+			wwlog.Warn("No container set for node %s", remoteNode.Id())
 		}
 	}
 
@@ -278,7 +274,7 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 				wwlog.ErrorExc(err, "")
 			}
 
-			wwlog.Send("%15s: %s", node.Id.Get(), stage_file)
+			wwlog.Send("%15s: %s", remoteNode.Id(), stage_file)
 
 		} else {
 			if rinfo.compress == "gz" {
@@ -296,24 +292,24 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 				w.WriteHeader(http.StatusNotFound)
 			}
 
-			err = sendFile(w, req, stage_file, node.Id.Get())
+			err = sendFile(w, req, stage_file, remoteNode.Id())
 			if err != nil {
 				wwlog.ErrorExc(err, "")
 				return
 			}
 		}
 
-		updateStatus(node.Id.Get(), status_stage, path.Base(stage_file), rinfo.ipaddr)
+		updateStatus(remoteNode.Id(), status_stage, path.Base(stage_file), rinfo.ipaddr)
 
 	} else if stage_file == "" {
 		w.WriteHeader(http.StatusBadRequest)
 		wwlog.Error("No resource selected")
-		updateStatus(node.Id.Get(), status_stage, "BAD_REQUEST", rinfo.ipaddr)
+		updateStatus(remoteNode.Id(), status_stage, "BAD_REQUEST", rinfo.ipaddr)
 
 	} else {
 		w.WriteHeader(http.StatusNotFound)
 		wwlog.Error("Not found: %s", stage_file)
-		updateStatus(node.Id.Get(), status_stage, "NOT_FOUND", rinfo.ipaddr)
+		updateStatus(remoteNode.Id(), status_stage, "NOT_FOUND", rinfo.ipaddr)
 	}
 
 }